# openapi.yaml
# UniPilot API — OpenAPI 3.0 (subset)
# Full reference: see API_OVERVIEW.md and server/routes/*.js

openapi: 3.0.3
info:
  title: UniPilot API
  description: Student dashboard and admin API. All protected routes require Authorization Bearer token.
  version: 1.0.0

servers:
  - url: /api
    description: API base path

paths:
  /health:
    get:
      summary: Health check
      tags: [Public]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }

  /ready:
    get:
      summary: Readiness (DB check)
      tags: [Public]
      responses:
        '200':
          description: Ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ready }
                  database: { type: string, example: connected }
        '503':
          description: Not ready

  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Returns access_token and user
        '400':
          description: Validation error
        '429':
          description: Rate limit

  /auth/register:
    post:
      summary: Register
      tags: [Auth]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                full_name: { type: string }
      responses:
        '200':
          description: Returns access_token and user
        '400':
          description: Validation error
        '429':
          description: Rate limit

  /auth/me:
    get:
      summary: Current user
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User object
        '401':
          description: Unauthorized

  /dashboard/summary:
    get:
      summary: Dashboard data (courses, GPA, tasks)
      tags: [Dashboard]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Summary object
        '401':
          description: Unauthorized

  /planner/daily:
    get:
      summary: Events and tasks for a date
      tags: [Planner]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Daily events and tasks
        '401':
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT